3rd Commit
Had to completely redo the project after my IntelliJ/Hadoop/Spark/sbt configuration completely broke. The project is in a Zip folder since the desktop app cannot locate the original, and I do want to risk cloning the repository and having to redo everything again. The Generator class in gone and replaced by a DatabaseManager class. There is also a Database Analysis class with functions for each query. The V2 menu system that Julio created is integrated and the first query will output to the console.