4th Commit
Everything is running fine in terms of the menu system and each of the queries  will output results to the terminal. However, the queries for Marketing Scenario 4 are giving me issues. After that issue is solves, then the last hurdle is incorporating the Visualization tool. Hopefully, our project will be finished by tonight!