6/5/2022 Commit
Incorporated the Login menu into the application. Also, created a Main class that can be utilized to run the entire application. Added several countries to the Generator for more variety when querying the data. A few things need to be cleaned up and hopefully we can further optimize. I am working on getting Tableau incorporated into IntelliJ; however, it is possible to just use the desktop app to create worksheets for each query.